一時退避

+++ app/controllers/public/pickups_controller.rb
@@ -5,14 +5,16 @@ class Public::PickupsController < ApplicationController
   def create
     Pickup.create(user_id: current_user.id, review_id: @review.id)
   end
-  
+
   def destroy
     pickup = Pickup.find_by(user_id: current_user.id, review_id: @review.id)
-    pickup.destroy
+    if pickup.prenset?
+       pickup.destroy
+    end
   end
-  
+
   private
-  
+
   def review_params
     @review = Review.find(params[:review_id])
   end
